prabhav/requirements.txt
flask
python-dotenv
openai
aiohttp
requests
gunicorn
pinecone-client[grpc]